Revert "divide some code related with html example"
